requirements.txt
asttokens==2.4.1
backcall==0.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
torch==2.4.1+cu118
torchaudio==2.4.1
torchvision==0.19.1+cu118
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.20.2
